refactor(auth): remove unused sign-in form data and schema
- Deleted the sign-in form data and schema file as it is no longer needed.
- Updated import paths in the sign-in form component to reflect the removal of the data file.
- Removed default values from the sign-in form configuration.
